subgraph(refactor): change mapping to camelCase
