+++ packages/subgraph/src/mappings/poolMetadataRegistry.ts
